Created tests into new test class
